Sensor Polling
Added a second ros::Timer which will 'poll' the sensor at some rate, so
that it is now simple to get the driver to respect the 'poll_rate'
parameter sent to it by DR/rosparam. 

As a backer to represent the possible disconnect between the rate that
the sensor actually updates and the rate that driver polls, added a
2-deep circular buffer to KVHDriverNode that stores measurements. The
poll timer callback pops from this buffer and does a filter update with
the measurement, or does an empty filter update if there is no new
measurements in the buffer. Right now the only thing that pushes to this
buffer is the test harness subscriber callback.

#2 #9