Constructor refactoring / OdometryFilter addition
Refactored the constructor for KVHDriverNode to call numerous
intializaiton methods that handle doing all of the various components of
node construction instead of it all being in a single monolithic
constructor. This improves code readability and debugability.

Also added construction of an odometry filter to KVHDriverNode for
testing  #1, although it is currently not used for anything