Added Odometry Filter Enable/Disable to dynamic_reconfigure
Added a boolean flag for enabling/disabiling odometry data
filtering/output to the dynamic_reconfigure parameters. Implemented a
levels callback for the flad in KVHDriverNode. Also refactored the name
of the IMU Filtering enable/disable paramter from filter to imu_filter.
Added a boolean flag to KVHDriverNode to store the state of odometry
filtering enable/disable.

#1 #13