Implemented Odometry Filtering in KVHDriverNode
Added a subscriber to the IMU topic published by the IMUFilter in
KVHDrivernode. Its callback extracts the state/covarance data from the
message, converts it to input/measurement data for OdometryFilter, and
then updates KVHDriverNode's OdometryFilter. Currently there is no
checking of the Enable/Disable flag for odom filtering.

Also refactored a few names in constants to reflect the fact that they
were exclusive to OdometryFilters and not all LinearFilters.

#1 #13